testing

+++ src/main/java/com/covid/countries/service/impl/CountryCovid19InfoServiceImpl.java
@@ -1,12 +1,9 @@
 package com.covid.countries.service.impl;
 
 import com.covid.countries.model.entities.CountryCovidInfo;
-//import com.covid.countries.model.modelgson.Country;
 import com.covid.countries.model.view.CountryCovidViewModel;
-//import com.covid.countries.model.view.Premium;
 import com.covid.countries.repository.CountriesCovidInfoRepository;
 import com.covid.countries.service.CountryCovid19InfoService;
-import com.google.gson.Gson;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
@@ -19,16 +16,14 @@ import java.util.Optional;
 public class CountryCovid19InfoServiceImpl implements CountryCovid19InfoService {
     private final CountriesCovidInfoRepository countriesCovidInfoRepository;
     private final ModelMapper modelMapper;
-//    private final Gson gson;
 
     public CountryCovid19InfoServiceImpl(CountriesCovidInfoRepository countriesCovidInfoRepository, ModelMapper modelMapper) {
         this.countriesCovidInfoRepository = countriesCovidInfoRepository;
         this.modelMapper = modelMapper;
-//        this.gson = gson;
     }
 
     @Override
-    public void processCovid19Data(String inputData) {
+    public List<CountryCovidInfo> processCovid19Data(String inputData) {
         List<CountryCovidInfo> countries = new ArrayList<>();
         if (inputData.contains("Countries=")) {
             String[] countriesInfo = inputData
@@ -59,36 +54,42 @@ public class CountryCovid19InfoServiceImpl implements CountryCovid19InfoService 
                                 .setPremium(splitInputData(currentCountryInfo[11]));
                 countries.add(countryCovidInfo);
             }
+        }
+        saveProcessedInputInfoInDb(countries);
+        return countries;
+    }
 
-            if (this.countriesCovidInfoRepository.count() == 0) {
-                this.countriesCovidInfoRepository.saveAll(countries);
-            } else {
-                List<CountryCovidInfo> allCountriesInDb = this.countriesCovidInfoRepository.findAll();
-                if (allCountriesInDb.size() == countries.size()) {
-                    for (int i = 0; i < allCountriesInDb.size(); i++) {
-                        CountryCovidInfo countryCovidInfo = allCountriesInDb.get(i);
-                        CountryCovidInfo countryInput = countries.get(i);
-                        countryCovidInfo
-                                .setID(countryInput.getID())
-                                .setCountry(countryInput.getCountry())
-                                .setCountryCode(countryInput.getCountryCode())
-                                .setSlug(countryInput.getSlug())
-                                .setNewConfirmed(countryInput.getNewConfirmed())
-                                .setTotalConfirmed(countryInput.getTotalConfirmed())
-                                .setNewDeaths(countryInput.getNewDeaths())
-                                .setTotalDeaths(countryInput.getTotalDeaths())
-                                .setNewRecovered(countryInput.getNewRecovered())
-                                .setTotalRecovered(countryInput.getTotalRecovered())
-                                .setDate(countryInput.getDate())
-                                .setPremium(countryInput.getPremium());
-                        this.countriesCovidInfoRepository.save(countryCovidInfo);
-                    }
-                } else {
-                    this.countriesCovidInfoRepository.deleteAll();
-                    this.countriesCovidInfoRepository.saveAll(countries);
+    @Override
+    public List<CountryCovidInfo> saveProcessedInputInfoInDb(List<CountryCovidInfo> countries) {
+        if (this.countriesCovidInfoRepository.count() == 0 && countries.size() != 0) {
+            this.countriesCovidInfoRepository.saveAll(countries);
+        } else {
+            List<CountryCovidInfo> allCountriesInDb = this.countriesCovidInfoRepository.findAll();
+            if (allCountriesInDb.size() == countries.size() && countries.size() != 0) {
+                for (int i = 0; i < allCountriesInDb.size(); i++) {
+                    CountryCovidInfo countryCovidInfo = allCountriesInDb.get(i);
+                    CountryCovidInfo countryInput = countries.get(i);
+                    countryCovidInfo
+                            .setID(countryInput.getID())
+                            .setCountry(countryInput.getCountry())
+                            .setCountryCode(countryInput.getCountryCode())
+                            .setSlug(countryInput.getSlug())
+                            .setNewConfirmed(countryInput.getNewConfirmed())
+                            .setTotalConfirmed(countryInput.getTotalConfirmed())
+                            .setNewDeaths(countryInput.getNewDeaths())
+                            .setTotalDeaths(countryInput.getTotalDeaths())
+                            .setNewRecovered(countryInput.getNewRecovered())
+                            .setTotalRecovered(countryInput.getTotalRecovered())
+                            .setDate(countryInput.getDate())
+                            .setPremium(countryInput.getPremium());
+                    this.countriesCovidInfoRepository.save(countryCovidInfo);
                 }
+            } else {
+                this.countriesCovidInfoRepository.deleteAll();
+                this.countriesCovidInfoRepository.saveAll(countries);
             }
         }
+        return this.countriesCovidInfoRepository.findAll();
     }
 
     @Override
@@ -97,10 +98,6 @@ public class CountryCovid19InfoServiceImpl implements CountryCovid19InfoService 
         if (countryByCountryCode.isPresent()) {
             CountryCovidInfo countryCovidInfo = countryByCountryCode.get();
             CountryCovidViewModel countryView = this.modelMapper.map(countryCovidInfo, CountryCovidViewModel.class);
-//            List<Premium> premiums = new ArrayList<>();
-//            Premium premium = new Premium();
-//            premium.setPremium(premium);
-//            countryView.setPremium(premium);
             return countryView;
         } else {
             return null;
